pyproject.toml
[project]
name = "predli-research-agent"
version = "2.0.0"
description = "Advanced agentic research system with parallel search, RAG, and template-based generation"
authors = [
    {name = "Predli Research Team"}
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    # Core LangGraph & LangChain (modern versions)
    "langgraph>=0.2.45",
    "langchain>=0.3.7",
    "langchain-openai>=0.2.9",
    "langchain-community>=0.3.5",
    "langchain-core>=0.3.15",
    
    # Vector Store & Embeddings
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.2.1",
    
    # Search APIs
    "tavily-python>=0.5.0",
    "arxiv>=2.1.3",
    "wikipedia>=1.4.0",
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.3",
    
    # FastAPI & Web
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.32.1",
    "python-multipart>=0.0.17",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.6.1",
    
    # PDF Generation
    "weasyprint>=62.3",
    "markdown2>=2.5.1",
    "jinja2>=3.1.4",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.27.2",
    "aiohttp>=3.11.7",
    "tenacity>=9.0.0",
    "tqdm>=4.67.1",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "black>=24.10.0",
    "ruff>=0.8.4",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
